requirements.txt
# Python-3.9.7
numpy==1.22.2
pandas==1.3.5
Pillow==10.0.0
streamlit==1.11.0
tensorflow==2.13.0